hal: Move backend specific functions and structs into their own module
