+++ src/backend/vulkan/src/lib.rs
@@ -374,7 +374,7 @@ impl hal::PhysicalDevice<Backend> for PhysicalDevice {
             .into_iter()
             .map(|(family, priorities)| {
                 let family_index = family.index;
-                let mut family_raw = queue::family::RawQueueGroup::new(family);
+                let mut family_raw = hal::backend::RawQueueGroup::new(family);
                 for id in 0 .. priorities.len() {
                     let queue_raw = unsafe {
                         device_arc.0.get_device_queue(family_index, id as _)
